+++ packages/iov-encoding/src/index.ts
@@ -8,3 +8,4 @@ export {
   JsonCompatibleArray,
   isJsonCompatibleArray,
 } from "./json";
+export { isNonNullObject, isUint8Array } from "./typechecks";