+++ packages/iov-encoding/types/index.d.ts
@@ -1,3 +1,4 @@
 export * from "./encoding";
 export { Int53, Uint32, Uint53, Uint64 } from "./integers";
 export { JsonCompatibleValue, isJsonCompatibleValue, JsonCompatibleDictionary, isJsonCompatibleDictionary, JsonCompatibleArray, isJsonCompatibleArray, } from "./json";
+export { isNonNullObject, isUint8Array } from "./typechecks";